requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pyyaml>=6.0
python-multipart>=0.0.9
watchdog>=4.0.0
